add SharedPreferences and coorect some issues
