+++ lib/app/widgets/not_found_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class NotFoundPage extends StatelessWidget {
+  const NotFoundPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: const Center(
+        child: Text('Not Found'),
+      ),
+    );
+  }
+}
