+++ lib/features/home/pages/bloc/home_bloc.dart
@@ -1,13 +1,13 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:project_video/app/constants.dart';
 import 'package:project_video/data/repositories/films_repository.dart';
-import 'package:project_video/features/home/pages/bloc/catalog_event.dart';
-import 'package:project_video/features/home/pages/bloc/catalog_state.dart';
+import 'package:project_video/features/home/pages/bloc/home_event.dart';
+import 'package:project_video/features/home/pages/bloc/home_state.dart';
 
-class CatalogBloc extends Bloc<CatalogEvent, CatalogState> {
+class HomeBloc extends Bloc<HomeEvent, HomeState> {
   final FilmsRepository repository;
 
-  CatalogBloc(this.repository) : super(const CatalogState()) {
+  HomeBloc(this.repository) : super(const HomeState()) {
     on<LoadDataEvent>(_onLoadData);
     on<SearchChangedEvent>(_onSearchChanged);
   }
@@ -19,13 +19,13 @@ class CatalogBloc extends Bloc<CatalogEvent, CatalogState> {
         : MovieQuery.initialQ;
   }
 
-  void _onSearchChanged(SearchChangedEvent event, Emitter<CatalogState> emit) {
+  void _onSearchChanged(SearchChangedEvent event, Emitter<HomeState> emit) {
     // search must be emitted before loading data
     emit(state.copyWith(search: event.search));
     emit(state.copyWith(data: repository.loadData(q: search)));
   }
 
-  void _onLoadData(LoadDataEvent event, Emitter<CatalogState> emit) {
+  void _onLoadData(LoadDataEvent event, Emitter<HomeState> emit) {
     emit(state.copyWith(data: repository.loadData(q: search)));
   }
 }
