+++ lib/features/home/pages/bloc/home_state.dart
@@ -1,14 +1,13 @@
 import 'package:equatable/equatable.dart';
 import 'package:project_video/app/models/home_model.dart';
 
-class CatalogState extends Equatable {
+class HomeState extends Equatable {
   final String? search;
   final Future<HomeModel?>? data;
 
-  const CatalogState({this.search, this.data});
+  const HomeState({this.search, this.data});
 
-  CatalogState copyWith({String? search, Future<HomeModel?>? data}) =>
-      CatalogState(
+  HomeState copyWith({String? search, Future<HomeModel?>? data}) => HomeState(
         search: search ?? this.search,
         data: data ?? this.data,
       );
