+++ lib/features/settings/bloc/setting_state.dart
@@ -1,0 +1,20 @@
+import 'package:equatable/equatable.dart';
+
+class SettingState extends Equatable {
+  final String? name;
+
+  const SettingState({
+    this.name,
+  });
+
+  @override
+  List<Object> get props => [name ?? ''];
+
+  SettingState copyWith({
+    String? name,
+  }) {
+    return SettingState(
+      name: name ?? this.name,
+    );
+  }
+}
