+++ pubspec.yaml
@@ -31,6 +31,8 @@ dependencies:
     sdk: flutter
   json_annotation: ^4.5.0
 
+  shared_preferences: ^2.0.15
+
   equatable: 2.0.3
   flutter_bloc: ^8.0.1
